feat(CategoriesController): use validateAccessLevel function

+++ src/controllers/CategoriesController.ts
@@ -6,12 +6,11 @@ import { queryPageSchema } from "../utils/queryPageSchema";
 import { GetManyCategoriesService } from "../services/categories/GetManyCategoriesService";
 import { CreateProductCategoryService } from "../services/categories/AddCategoryToProductService";
 import { DeleteCategoryFromExistingProductService } from "../services/categories/DeleteCategoryFromExistingProductService";
-import { JwtPayload } from "jsonwebtoken";
+import { validateAccessLevel } from "../utils/validateAccessLevel";
 
 export class CategoriesController {
   create = async (req: Request, res: Response) => {
-    const payload: JwtPayload = (req as any).jwtPayload;
-    if (Number(payload.userAccessLevel) < 2) res.status(403).json({ message: "Forbidden access." });
+    validateAccessLevel(2, req);
     const data = req.body;
     const category = await new CreateCategoryService().execute(data);
     return res.status(200).json({
@@ -34,8 +33,7 @@ export class CategoriesController {
   };
 
   deleteOne = async (req: Request, res: Response) => {
-    const payload: JwtPayload = (req as any).jwtPayload;
-    if (Number(payload.userAccessLevel) < 2) res.status(403).json({ message: "Forbidden access." });
+    validateAccessLevel(2, req);
     const categoryID = req.params.categoryID;
     await new DeleteOneCategoryService().execute(categoryID);
     return res.status(200).json({
@@ -44,16 +42,14 @@ export class CategoriesController {
   };
 
   addCategoryToExistingProduct = async (req: Request, res: Response) => {
-    const payload: JwtPayload = (req as any).jwtPayload;
-    if (Number(payload.userAccessLevel) < 2) res.status(403).json({ message: "Forbidden access." });
+    validateAccessLevel(2, req);
     const { productID, categoryID } = req.body;
     await new CreateProductCategoryService().execute(productID, categoryID);
     return res.json({ message: "Category successfully added to product." });
   };
 
   deleteCategoryFromExistingProduct = async (req: Request, res: Response) => {
-    const payload: JwtPayload = (req as any).jwtPayload;
-    if (Number(payload.userAccessLevel) < 2) res.status(403).json({ message: "Forbidden access." });
+    validateAccessLevel(2, req);
     const { productID, categoryID } = req.body;
     await new DeleteCategoryFromExistingProductService().execute(productID, categoryID);
     res.status(200).json({ message: "Category successfully removed." });
